requirements.txt
Django>=5.2,<6
djangorestframework>=3.14
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
psycopg[binary]>=3.2
Faker>=24.0
